+++ keyboards/bbrfkr/dynamis/keymaps/iso/keymap.c
@@ -57,14 +57,3 @@ const uint16_t PROGMEM keymaps[][MATRIX_ROWS][MATRIX_COLS] = {
       KC_TRNS, KC_TRNS, KC_TRNS, KC_TRNS,          KC_TRNS,                   KC_TRNS, KC_TRNS, KC_TRNS, KC_TRNS,          KC_TRNS, KC_TRNS, KC_TRNS \
   )
 };
-
-bool encoder_update_user(uint8_t index, bool clockwise) {
-   if (index == 0) { /* First encoder */
-      if (clockwise) {
-         tap_code(KC_WH_D);
-      } else {
-         tap_code(KC_WH_U);
-      }
-   }
-   return false;
-}
