+++ keyboards/bbrfkr/dynamis/readme.md
@@ -1,5 +1,6 @@
 # dynamis
 
+## general information
 ![dynamis](https://i.imgur.com/q2C3D2u.jpg)
 
 Gasket mounted 65% row staggered keyboard integrated a trackball
@@ -7,6 +8,15 @@ Gasket mounted 65% row staggered keyboard integrated a trackball
 * Keyboard Maintainer: [bbrfkr](https://github.com/bbrfkr/)
 * Hardware Supported: dynamis PCBs
 * Hardware Availability: http://github.com/bbrfkr/keyboards
+
+## bootloader
+
+Enter the bootloader in 2 ways:
+
+* **Physical reset button**: Briefly press the button on the back of the PCB - some may have pads you must short instead
+* **Keycode in layout**: Press the key mapped to `RESET` if it is available
+
+## build and flash firmware
 
 Make example for this keyboard (after setting up your build environment):
